reworked rendering pipeline, changed coding style :)

+++ src/globals.c
@@ -39,9 +39,9 @@ float r_cosfov2;
 float r_tanfov;
 int r_eyecluster;
 
-double sv_frametime = 0.0;
-double cl_frametime = 0.0;
-double shadertime = 0.0;
+float sv_frametime = 0.0;
+float cl_frametime = 0.0;
+float shadertime = 0.0;
 
 uint_t *r_lightmaptex;
 int r_numlightmaptex;
